refactor: removed unnecessary `eventService` into its own function `getEvents()` + added documentation for `searchEvents()` and `sortEvents()`
